Move workspace folder functions to separate file
